.github/workflows/scripts/requirements.txt
zabbix-utils>=2.0.0
tabulate
ruamel.yaml
lxml